// packages/mods/flight-planning/src/features/expert-settings/types/routeEnergySettings.ts
export type RouteEnergySettings = {
    windSpeed?: number;
    windDirection?: number;
    takeoffMass?: number;
    batteryCycles?: number;
    airDensity?: number;
    pilotEfficiencySafetyBuffer?: number;
    transitionAltitude?: number;
    airspeedCsfl?: number;
    descendAngleCsfl?: number;
    climbTakeoff?: number;
    climbTouchDown?: number;
    maxClimbRate?: number;
    maxDescendRate?: number;
    loiterAltitude?: number;
    loiterTime?: number;
    loiterTimeCsfl?: number;
    batteryCapacityOrig?: number;
    unusableEnergy?: number;
    finalReserve?: number;
    additionalEnergy?: number;
    extraEnergy?: number;
    discretionaryEnergy?: number;
    contingencyEnergy?: number;
    goAroundEnergy?: number;
    integrationTimeStepsSeconds?: number;
    ambientTemperature?: number;
    csflEnergyOverhead?: number;
    pdmTwoEfficiencyLoss?: number;
    maxTimeToCsfl?: number;
    departureTaxiSpeed?: number;
    departureTaxiDistance?: number;
    arrivalTaxiSpeed?: number;
    arrivalTaxiDistance?: number;
    horizontalObstacleClearance?: number;
    verticalObstacleClearance?: number;
};